Extract Zcash into walletkit-chain-zcash module
Moves all SDK-importing code (adapter cluster, birthday provider, send
and shield flows, network settings, address-type select page, validators
with the rust-backend static path, plugin) into a new library module
with the zcash sdk and backend dependencies. Kit-free stays in core:
ZcashLightWalletEndpointManager and endpoint storage, backup DTOs,
ILocalStorage zcash fields, BalanceData.unshielded, and the shielding
transaction record whose ShieldDirection enum moves into the record
file. ISendZcashAdapter relocates next to its consumers.
