+++ settings.gradle.kts
@@ -22,6 +22,7 @@ include(":app")
 include(":walletkit")
 include(":walletkit-chain-zano")
 include(":walletkit-chain-monero")
+include(":walletkit-chain-zcash")
 include(":components:icons")
 include(":components:chartview")
 include(":subscriptions-core")
