+++ walletkit-chain-zcash/build.gradle.kts
@@ -1,0 +1,47 @@
+plugins {
+    alias(libs.plugins.android.library)
+    alias(libs.plugins.kotlin.compose)
+    alias(libs.plugins.jetbrains.kotlin.serialization)
+}
+
+kotlin {
+    compilerOptions {
+        jvmTarget.set(org.jetbrains.kotlin.gradle.dsl.JvmTarget.JVM_17)
+    }
+}
+
+android {
+    namespace = "io.horizontalsystems.walletkit.chain.zcash"
+    compileSdk = libs.versions.compileSdk.get().toInt()
+
+    defaultConfig {
+        minSdk = libs.versions.minSdk.get().toInt()
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+
+    buildFeatures {
+        compose = true
+    }
+}
+
+dependencies {
+    api(project(":walletkit"))
+
+    api(libs.zcash.android.sdk)
+    implementation(libs.zcash.android.backend)
+
+    testImplementation(libs.junit)
+}
+
+// Forward -PupdateParityFixture=true to the test JVM so ChainBehaviorParityTest can
+// regenerate its golden fixture (see docs/Walletkit-Modularization-Plan.md).
+tasks.withType<Test>().configureEach {
+    systemProperty(
+        "updateParityFixture",
+        providers.gradleProperty("updateParityFixture").getOrElse("false")
+    )
+}
