+++ walletkit/build.gradle.kts
@@ -183,8 +183,6 @@ dependencies {
     api(libs.kit.solana)
     api(libs.kit.thorchain)
     api(libs.kit.tron)
-    api(libs.zcash.android.sdk)
-    implementation(libs.zcash.android.backend)
 
     // BouncyCastle
     api(libs.bouncycastle)
